chore: move services and models into pkg
