+++ handlers/Index.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"time"
 
-	"jirku.sk/mcmamina/models"
+	"jirku.sk/mcmamina/pkg/models"
 	"jirku.sk/mcmamina/template/components"
 	"jirku.sk/mcmamina/template/pages"
 )
