+++ template/components/twoWeeksCalendar.templ
@@ -2,7 +2,7 @@ package components
 
 import (
 	"time"
-	"jirku.sk/mcmamina/models"
+	"jirku.sk/mcmamina/pkg/models"
 )
 
 templ TwoWeeksCalendar(days []models.Day, today time.Time) {
