+++ template/pages/Index.templ
@@ -4,7 +4,7 @@ import (
 	"time"
 	"jirku.sk/mcmamina/template/layout"
 	"jirku.sk/mcmamina/template/components"
-	"jirku.sk/mcmamina/models"
+	"jirku.sk/mcmamina/pkg/models"
 )
 
 type IndexPageDto struct {
