+++ template/pages/SupportedUs.templ
@@ -4,7 +4,7 @@ package pages
 import (
 	"jirku.sk/mcmamina/template/layout"
 	"jirku.sk/mcmamina/template/components"
-	"jirku.sk/mcmamina/models"
+	"jirku.sk/mcmamina/pkg/models"
 	"fmt"
 )
